requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
